+++ http/src/http.ml
@@ -162,14 +162,20 @@ module Header = struct
         add_multi h k xs
 
   let move_to_front t hdr_name =
-    match t with
-    | (k, _) :: _ when caseless_equal k hdr_name -> t
-    | _ -> (
-        match get t hdr_name with
-        | Some v ->
-            let headers = remove t hdr_name in
-            add headers hdr_name v
-        | None -> t)
+    (* Headers are manipulated in reverse order for convenience, so we
+       need to reverse them, prepend what we need, and reverse again *)
+    let t = List.rev t in
+    let t =
+      match t with
+      | (k, _) :: _ when caseless_equal k hdr_name -> t
+      | _ -> (
+          match get t hdr_name with
+          | Some v ->
+              let headers = remove t hdr_name in
+              add headers hdr_name v
+          | None -> t)
+    in
+    List.rev t
 
   let map (f : string -> string -> string) (h : t) : t =
     List.map
@@ -178,6 +184,9 @@ module Header = struct
         (k, vs'))
       h
 
+  let iter_ord (f : string -> string -> unit) (h : t) : unit =
+    List.iter (fun (k, v) -> f k v) (List.rev h)
+  
   let iter (f : string -> string -> unit) (h : t) : unit =
     List.iter (fun (k, v) -> f k v) h
 
@@ -349,7 +358,7 @@ module Header = struct
 
   module Private = struct
     let caseless_equal = caseless_equal
-    let first = first
+    let first l = first (List.rev l)
     let move_to_front = move_to_front
   end
 end